trading_bots/requirements.txt
requests>=2.31.0

bitunix-api>=1.0.0

python-dotenv>=1.0.0
